+++ src/components/NavBar.tsx
@@ -34,12 +34,6 @@ export default function NavBar(props: navProps) {
             orientation="vertical"
             flexItem
           />
-          <Link to={`/wishlist/${props.user}`}>Wishlist</Link>
-          <Divider
-            sx={{ background: '#334195', marginLeft: 1, marginRight: 1 }}
-            orientation="vertical"
-            flexItem
-          />
           <Link to="/search">Add Games</Link>
           <Divider
             sx={{ background: '#334195', marginLeft: 1, marginRight: 1 }}
